3437 - refactor odp comments with datagrid and new css classes
